MAINT Move classes and functions into utils.base
